+++ spec/integration/puppet_x/puppetlabs/scheduled_task/v2adapter_spec.rb
@@ -64,6 +64,26 @@ end
 # and read back correctly
 describe "PuppetX::PuppetLabs::ScheduledTask::V2Adapter", :if => Puppet.features.microsoft_windows? do
   subject = PuppetX::PuppetLabs::ScheduledTask::V2Adapter
+  V2 = PuppetX::PuppetLabs::ScheduledTask::Trigger::V2
+
+  context "should ignore unknown Trigger types" do
+    [
+      { :ole_type => 'IIdleTrigger', :Type => V2::Type::TASK_TRIGGER_IDLE, },
+      { :ole_type => 'IRegistrationTrigger', :Type => V2::Type::TASK_TRIGGER_REGISTRATION, },
+      { :ole_type => 'ILogonTrigger', :Type => V2::Type::TASK_TRIGGER_LOGON, },
+      { :ole_type => 'ISessionStateChangeTrigger', :Type => V2::Type::TASK_TRIGGER_SESSION_STATE_CHANGE, },
+      { :ole_type => 'IEventTrigger', :Type => V2::Type::TASK_TRIGGER_EVENT, },
+    ].each do |trigger_details|
+      it "by returning nil for a #{trigger_details[:ole_type]} instance" do
+        task_object = subject.new('foo')
+        # guarantee task not saved to system
+        task_object.stubs(:save)
+        PuppetX::PuppetLabs::ScheduledTask::TaskScheduler2.expects(:trigger).returns(stub(trigger_details))
+
+        expect(task_object.trigger(0)).to be_nil
+      end
+    end
+  end
 
   context "should be able to create trigger" do
     before(:all) do